Renames the test
