+++ springdoc-openapi-webmvc-core/src/test/java/test/org/springdoc/api/v30/app190/ConcreteSubclassFromGeneric.java
@@ -1,4 +1,4 @@
-package test.org.springdoc.api.v30.app189;
+package test.org.springdoc.api.v30.app190;
 
 public class ConcreteSubclassFromGeneric extends SimpleGeneric<MyData> {
     private String topName;
