+++ springdoc-openapi-webmvc-core/src/test/java/test/org/springdoc/api/v30/app190/HelloController.java
@@ -20,7 +20,7 @@
  *
  */
 
-package test.org.springdoc.api.v30.app189;
+package test.org.springdoc.api.v30.app190;
 
 
 import org.springdoc.api.annotations.ParameterObject;
