+++ springdoc-openapi-webmvc-core/src/test/resources/results/3.0.1/app190.json
